Extract shared crate
