+++ Cargo.toml
@@ -1,14 +1,16 @@
 [workspace]
 resolver = "2"
-members = [
-  "crates/*"
-]
+members = ["crates/*"]
 
 [workspace.dependencies]
-hotshot = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77", features = ["dependency-tasks"] }
+hotshot = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77", features = [
+    "dependency-tasks",
+] }
 hotshot-builder-api = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77" }
 hotshot-events-service = { git = "https://github.com/EspressoSystems/hotshot-events-service.git", tag = "0.1.49" }
-hotshot-task-impls = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77", features = ["dependency-tasks"] }
+hotshot-task-impls = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77", features = [
+    "dependency-tasks",
+] }
 hotshot-types = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77" }
 hotshot-example-types = { git = "https://github.com/EspressoSystems/HotShot.git", tag = "0.5.77" }
 
@@ -48,4 +50,5 @@ edition = "2021"
 [workspace.lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = [
     'cfg(async_executor_impl, values("async-std", "tokio"))',
+    'cfg(async_channel_impl, values("async-std", "tokio"))',
 ] }
