+++ crates/legacy/Cargo.toml
@@ -4,6 +4,7 @@ version = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
+marketplace-builder-shared = { path = "../shared" }
 
 anyhow = { workspace = true }
 async-broadcast = { workspace = true }
