+++ crates/marketplace/Cargo.toml
@@ -4,6 +4,7 @@ version = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
+marketplace-builder-shared = { path = "../shared" }
 
 anyhow = "1"
 async-broadcast = "0.7"
