+++ crates/shared/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "marketplace-builder-shared"
+version = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+async-broadcast = { workspace = true }
+hotshot = { workspace = true }
+hotshot-builder-api = { workspace = true }
+hotshot-events-service = { workspace = true }
+hotshot-task-impls = { workspace = true }
+hotshot-types = { workspace = true }
+
+async-trait = { workspace = true }
+committable = { workspace = true }
+either = { workspace = true }
+futures = { workspace = true }
+hex = { workspace = true }
+surf-disco = { workspace = true }
+tracing = { workspace = true }
+url = { workspace = true }
+vbs = { workspace = true }
+
+[lints]
+workspace = true
