+++ crates/shared/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod block;
+#[cfg(test)]
+pub mod testing;
+pub mod utils;
